refactor: move auth-related mutation hooks into auth folder
